mv'ed function ccnl_interest_new
The commit removes dead code from ccnl-interest.c namely a function
with the following defintion:

  struct ccnl_interest_s*
    ccnl_interest_new(struct ccnl_face_s *from, struct ccnl_pkt_s **pkt)

and moves a function with the name, but different parameters from
ccnl-relay.c to ccnl-interest.h. The definition is as follows

  struct ccnl_interest_s*
    ccnl_interest_new(struct ccnl_relay_s *ccnl, struct ccnl_face_s *from,
       struct ccnl_pkt_s **pkt);
